Begun work on game phases and transitions.
