+++ src/civilization/mod.rs
@@ -1,1 +1,2 @@
 pub mod civ;
+pub mod population_expansion;
