+++ src/console/commands.rs
@@ -2,7 +2,8 @@ use bevy::app::{App, Plugin};
 use bevy::prelude::{Children, Entity, EventWriter, Parent, Query, With};
 use bevy_console::{AddConsoleCommand, ConsoleCommand, ConsoleConfiguration, ConsolePlugin};
 use clap::Parser;
-use crate::civilization::civ::{CheckPopulationExpansionEligibilityEvent, MoveTokensFromStockToAreaCommand, Population, StartArea};
+use crate::civilization::civ::{MoveTokensFromStockToAreaCommand, Population, StartArea};
+use crate::civilization::population_expansion::CheckPopulationExpansionEligibilityEvent;
 use crate::player::Player;
 
 pub struct CommandsPlugin;
